+++ feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/mapper/response/AfternoteListResponseToListPage.kt
@@ -2,7 +2,7 @@ package com.afternote.feature.afternote.data.mapper.response
 
 import com.afternote.feature.afternote.data.dto.AfternoteListResponse
 import com.afternote.feature.afternote.data.mapper.toDomainList
-import com.afternote.feature.afternote.domain.model.ListPage
+import com.afternote.feature.afternote.domain.model.author.ListPage
 
 fun AfternoteListResponse.toListPage() =
     ListPage(
