+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/repository/AfternoteRepository.kt
@@ -1,11 +1,11 @@
 package com.afternote.feature.afternote.domain.repository
 
-import com.afternote.feature.afternote.domain.model.Detail
-import com.afternote.feature.afternote.domain.model.ListPage
 import com.afternote.feature.afternote.domain.model.author.AfternoteUpdatePayload
 import com.afternote.feature.afternote.domain.model.author.CreateGalleryPayload
 import com.afternote.feature.afternote.domain.model.author.CreatePlaylistPayload
 import com.afternote.feature.afternote.domain.model.author.CreateSocialPayload
+import com.afternote.feature.afternote.domain.model.author.Detail
+import com.afternote.feature.afternote.domain.model.author.ListPage
 import kotlinx.coroutines.flow.StateFlow
 
 /**
